Refactor oauthEnabledClient and dependencies in Typescript
