+++ backend/@types/superagent/index.d.ts
@@ -1,0 +1,8 @@
+export default {}
+
+declare module 'superagent' {
+  interface Response {
+    req: superagent.Request
+    request: superagent.Request
+  }
+}
