+++ backend/apis.js
@@ -1,12 +1,12 @@
 const config = require('./config').default
-const clientFactory = require('./api/oauthEnabledClient')
 const { oauthApiFactory } = require('./api/oauthApi')
 const { manageUsersApiFactory } = require('./api/manageUsersApi')
 const { tokenVerificationApiFactory } = require('./api/tokenVerificationApi')
 const { nomisUsersAndRolesFactory } = require('./api/nomisUsersAndRolesApi')
+const { oauthEnabledClientFactory } = require('./api/oauthEnabledClient')
 
 const oauthApi = oauthApiFactory(
-  clientFactory({
+  oauthEnabledClientFactory({
     baseUrl: config.apis.hmppsAuth.url,
     timeout: config.apis.hmppsAuth.timeoutSeconds * 1000,
   }),
@@ -14,21 +14,21 @@ const oauthApi = oauthApiFactory(
 )
 
 const manageUsersApi = manageUsersApiFactory(
-  clientFactory({
+  oauthEnabledClientFactory({
     baseUrl: config.apis.manageUsers.url,
     timeout: config.apis.manageUsers.timeout.response,
   }),
 )
 
 const tokenVerificationApi = tokenVerificationApiFactory(
-  clientFactory({
+  oauthEnabledClientFactory({
     baseUrl: config.apis.tokenVerification.url,
     timeout: config.apis.tokenVerification.timeoutSeconds * 1000,
   }),
 )
 
 const nomisUsersAndRolesApi = nomisUsersAndRolesFactory(
-  clientFactory({
+  oauthEnabledClientFactory({
     baseUrl: config.apis.nomisUsersAndRoles.url,
     timeout: config.apis.nomisUsersAndRoles.timeoutSeconds * 1000,
   }),
